refactor: Refactor E2E alerts
- Move client related utils to ducks/client/utils
- Move serviceUtils outside the test file
