+++ test/e2e/alerts/alerts.e2e.spec.js
@@ -1,8 +1,6 @@
 /* eslint-disable no-console */
 
-import { spawnSync } from 'child_process'
 import pick from 'lodash/pick'
-import pickBy from 'lodash/pickBy'
 import omit from 'lodash/omit'
 
 import { createClientInteractive } from 'cozy-client/dist/cli'
@@ -12,35 +10,21 @@ import {
   SETTINGS_DOCTYPE,
   GROUP_DOCTYPE,
   BILLS_DOCTYPE
-} from '../../src/doctypes'
-import { importData } from './dataUtils'
+} from 'src/doctypes'
 import Mailhog from 'mailhog'
-import MockServer from './mock-server'
+import MockServer from '../mock-server'
 import scenarios from './scenarios'
 import fs from 'fs'
+import {
+  dropDoctype,
+  importACHData,
+  revokeOtherOAuthClientsForSoftwareId
+} from 'ducks/client/utils'
+import { runService } from 'test/e2e/serviceUtils'
 
 const SOFTWARE_ID = 'banks.alerts-e2e'
 
 jest.setTimeout(10 * 1000)
-
-const revokeOtherOAuthClientsForSoftwareId = async (client, softwareID) => {
-  const { data: clients } = await client.stackClient.fetchJSON(
-    'GET',
-    `/settings/clients`
-  )
-  const currentOAuthClientId = client.stackClient.oauthOptions.clientID
-  const otherOAuthClients = clients.filter(
-    oauthClient =>
-      oauthClient.attributes.software_id === softwareID &&
-      oauthClient.id !== currentOAuthClientId
-  )
-  for (let oauthClient of otherOAuthClients) {
-    await client.stackClient.fetchJSON(
-      'DELETE',
-      `/settings/clients/${oauthClient.id}`
-    )
-  }
-}
 
 const decodeEmail = (mailhog, attrs) =>
   attrs
@@ -49,33 +33,6 @@ const decodeEmail = (mailhog, attrs) =>
         subject: attrs.subject.replace(/_/g, ' ')
       }
     : attrs
-
-const runService = async options => {
-  const env = {
-    ...process.env,
-    IS_TESTING: 'test'
-  }
-  const processOptions = pickBy(
-    {
-      stdio: options.showOutput ? 'inherit' : undefined,
-      env
-    },
-    Boolean
-  )
-  const res = spawnSync(
-    'node',
-    ['build/onOperationOrBillCreate'],
-    processOptions
-  )
-
-  if (res.status !== 0) {
-    console.error(`Error: onOperationOrBillCreate exited with 1.`)
-    if (!options.showOutput) {
-      console.error(`Re-run with -v to see its output.`)
-    }
-    throw new Error('Error while running onOperationOrBillCreate')
-  }
-}
 
 const checkEmailForScenario = async (mailhog, scenario) => {
   const latestMessages = (await mailhog.messages(0, 1)).items
@@ -106,7 +63,7 @@ const checkPushForScenario = async (pushServer, scenario) => {
 }
 
 const runScenario = async (client, scenario, options) => {
-  await importData(client, scenario.data)
+  await importACHData(client, scenario.data)
 
   if (options.mailhog) {
     await options.mailhog.deleteAll()
@@ -115,7 +72,15 @@ const runScenario = async (client, scenario, options) => {
     options.pushServer.clearRequests()
   }
 
-  await runService(options)
+  try {
+    await runService('onOperationOrBillCreate', options)
+  } catch (e) {
+    console.error(e.message)
+    if (!options.showOutput) {
+      console.error(`Re-run with -v to see its output.`)
+    }
+    throw e
+  }
 
   if (options.mailhog) {
     const emailMatch = await checkEmailForScenario(options.mailhog, scenario)
@@ -134,13 +99,7 @@ const cleanupDatabase = async client => {
     GROUP_DOCTYPE,
     BILLS_DOCTYPE
   ]) {
-    const col = client.collection(doctype)
-    const { data: docs } = await col.getAll()
-    if (docs.length > 0) {
-      // The omit for _type can be removed when the following PR is resolved
-      // https://github.com/cozy/cozy-client/pull/597
-      await col.destroyAll(docs.map(doc => omit(doc, '_type')))
-    }
+    await dropDoctype(client, doctype)
   }
 }
